send_analytics_bot/requirements.txt
aiofiles==23.1.0
aiogoogle==5.7.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
google==3.0.0
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
gspread==6.1.0
httplib2==0.22.0
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
oauth2client==4.1.3
oauthlib==3.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyaes==1.6.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.3.0
pydantic-settings==2.2.1
pydantic_core==2.6.3
pyparsing==3.1.2
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
StrEnum==0.4.15
tonyg-rfc3339==0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
